small type refactor
